Fix _win_mkfifo prototype to match the declaration
git-svn-id: https://svn.code.sf.net/p/plibc/code/trunk/plibc@156 1596915d-9e13-0410-ab7f-96e59408e0f0

+++ src/include/plibc.h
@@ -522,7 +522,7 @@ int _win_ftruncate(int fildes, off_t length);
 int _win_truncate(const char *fname, int distance);
 int _win_kill(pid_t pid, int sig);
 int _win_pipe(int *phandles);
-intptr_t _win_mkfifo(const char *path, mode_t mode);
+int _win_mkfifo(const char *path, mode_t mode);
 int _win_rmdir(const char *path);
 int _win_access( const char *path, int mode );
 int _win_chmod(const char *filename, int pmode);
